submit_example/model/requirements.txt
opencv-python
Pillow
numpy
tifffile